# openapi.yaml
openapi: 3.0.0
info:
  title: reparafacil API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3013/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /reparacion:
    get:
      tags:
        - Reparacion
      summary: Listar todos los Reparacions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Reparacions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Reparacion
      summary: Crear Reparacion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Reparacion creado
  /reparacion/{id}:
    get:
      tags:
        - Reparacion
      summary: Obtener Reparacion por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reparacion encontrado
        '404':
          description: Reparacion no encontrado
    patch:
      tags:
        - Reparacion
      summary: Actualizar Reparacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Reparacion actualizado
    delete:
      tags:
        - Reparacion
      summary: Eliminar Reparacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reparacion eliminado
  /reparacion/{id}/upload-image:
    post:
      tags:
        - Reparacion
      summary: Subir imagen a Reparacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /agenda:
    get:
      tags:
        - Agenda
      summary: Listar todos los Agendas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Agendas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Agenda
      summary: Crear Agenda
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Agenda creado
  /agenda/{id}:
    get:
      tags:
        - Agenda
      summary: Obtener Agenda por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agenda encontrado
        '404':
          description: Agenda no encontrado
    patch:
      tags:
        - Agenda
      summary: Actualizar Agenda
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Agenda actualizado
    delete:
      tags:
        - Agenda
      summary: Eliminar Agenda
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agenda eliminado
  /agenda/{id}/upload-image:
    post:
      tags:
        - Agenda
      summary: Subir imagen a Agenda
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /garantia:
    get:
      tags:
        - Garantia
      summary: Listar todos los Garantias
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Garantias
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Garantia
      summary: Crear Garantia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Garantia creado
  /garantia/{id}:
    get:
      tags:
        - Garantia
      summary: Obtener Garantia por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Garantia encontrado
        '404':
          description: Garantia no encontrado
    patch:
      tags:
        - Garantia
      summary: Actualizar Garantia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Garantia actualizado
    delete:
      tags:
        - Garantia
      summary: Eliminar Garantia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Garantia eliminado
  /garantia/{id}/upload-image:
    post:
      tags:
        - Garantia
      summary: Subir imagen a Garantia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida